Fully working imu node
