+++ ROS_ws/src/imu/CMakeLists.txt
@@ -19,7 +19,7 @@ find_package(catkin REQUIRED)
 ## Uncomment this if the package has a setup.py. This macro ensures
 ## modules and global scripts declared therein get installed
 ## See http://ros.org/doc/api/catkin/html/user_guide/setup_dot_py.html
-#catkin_python_setup()
+catkin_python_setup()
 
 ################################################
 ## Declare ROS messages, services and actions ##
@@ -158,12 +158,10 @@ include_directories(
 
 ## Mark executable scripts (Python etc.) for installation
 ## in contrast to setup.py, you can choose the destination
-catkin_install_python(PROGRAMS  
-scripts/imu_node.py
-scripts/computation.py
-scripts/lsm6ds0.py
-  DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
-)
+#catkin_install_python(PROGRAMS  
+#scripts/run_imu.py
+#  DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
+#)
