+++ ROS_ws/src/imu/launch/imu.launch
@@ -1,3 +1,3 @@
 <launch>
-	<node pkg="imu" type="imu_node.py" name="imu" output="screen"></node>
+	<node pkg="imu" type="imu.py" name="imu" output="screen"></node>
 </launch>
