+++ ROS_ws/src/imu/src/__init__.py
@@ -1,0 +1,1 @@
+print("imu_package imported")
